+++ relay-redis/src/noop.rs
@@ -2,11 +2,11 @@ use thiserror::Error;
 
 use crate::config::RedisConfigOptions;
 
-/// This is an unconstructable type to make `Option<RedisPool>` zero-sized.
+/// This is an unconstructable type to make [`Option<RedisClient>`] zero-sized.
 #[derive(Clone, Debug)]
-pub struct RedisPool;
+pub struct RedisClient;
 
-/// An error returned from `RedisPool`.
+/// An error returned from [`RedisClient`].
 #[derive(Debug, Error)]
 #[error("unreachable")]
 pub enum RedisError {
@@ -14,10 +14,10 @@ pub enum RedisError {
     Configuration,
 }
 
-impl RedisPool {
-    /// Creates a `RedisPool` in cluster configuration.
+impl RedisClient {
+    /// Creates [`RedisClient`] in cluster configuration.
     ///
-    /// Always returns `Ok(Self)`.
+    /// Always returns [`Ok(Self)`].
     pub fn cluster<'a>(
         _servers: impl IntoIterator<Item = &'a str>,
         _opts: RedisConfigOptions,
@@ -25,41 +25,41 @@ impl RedisPool {
         Ok(Self)
     }
 
-    /// Creates a [`RedisPool`] in multi write configuration.
+    /// Creates a [`RedisClient`] in multi write configuration.
     ///
-    /// Always returns `Ok(Self)`.
+    /// Always returns [`Ok(Self)`].
     pub fn multi_write(
-        _primary: RedisPool,
-        _secondaries: Vec<RedisPool>,
+        _primary: RedisClient,
+        _secondaries: Vec<RedisClient>,
     ) -> Result<Self, RedisError> {
         Ok(Self)
     }
 
-    /// Creates a `RedisPool` in single-node configuration.
+    /// Creates a [`RedisClient`] in single-node configuration.
     ///
-    /// Always returns `Ok(Self)`.
+    /// Always returns [`Ok(Self)`].
     pub fn single(_server: &str, _opts: RedisConfigOptions) -> Result<Self, RedisError> {
         Ok(Self)
     }
 }
 
-/// The various [`RedisPool`]s used within Relay.
+/// The various [`RedisClient`]s used within Relay.
 #[derive(Debug, Clone)]
-pub struct RedisPools {
-    /// The pool used for project configurations
-    pub project_configs: RedisPool,
-    /// The pool used for cardinality limits.
-    pub cardinality: RedisPool,
-    /// The pool used for rate limiting/quotas.
-    pub quotas: RedisPool,
+pub struct RedisClients {
+    /// The client used for project configurations
+    pub project_configs: RedisClient,
+    /// The client used for cardinality limits.
+    pub cardinality: RedisClient,
+    /// The client used for rate limiting/quotas.
+    pub quotas: RedisClient,
 }
 
-/// Noop type of the `AsyncRedisPool`
+/// Noop type of the [`AsyncRedisClient`].
 #[derive(Debug, Clone)]
-pub struct AsyncRedisPool;
+pub struct AsyncRedisClient;
 
-impl AsyncRedisPool {
-    /// Creates a [`AsyncRedisPool`] in cluster configration.
+impl AsyncRedisClient {
+    /// Creates a [`AsyncRedisClient`] in cluster configration.
     ///
     /// Always returns `Ok(Self)`
     pub async fn cluster<'a>(
@@ -69,7 +69,7 @@ impl AsyncRedisPool {
         Ok(Self)
     }
 
-    /// Creates a [`AsyncRedisPool`] in single-node configuration.
+    /// Creates a [`AsyncRedisClient`] in single-node configuration.
     ///
     /// Always returns `Ok(Self)`
     pub async fn single(_server: &str, _opts: &RedisConfigOptions) -> Result<Self, RedisError> {