+++ relay-sampling/src/evaluation.rs
@@ -79,7 +79,7 @@ impl ReservoirEvaluator<'_> {
         client: &AsyncRedisClient,
         rule_expiry: Option<&DateTime<Utc>>,
     ) -> anyhow::Result<i64> {
-        let mut connection = client.get_connection();
+        let mut connection = client.get_connection().await?;
 
         let val = redis_sampling::increment_redis_reservoir_count(&mut connection, key).await?;
         redis_sampling::set_redis_expiry(&mut connection, key, rule_expiry).await?;